Break up store.js into multiple files
