+++ JTime-rest/client/jtime-website/src/store/toast.js
@@ -1,0 +1,13 @@
+export default {
+  state: {
+    toastQueue: [] // Queue of strings to be shown as toasts
+  },
+  mutations: {
+    toast (state, msg) {
+      state.toastQueue.push(msg)
+    },
+    removeOneToast (state) {
+      state.toastQueue.shift()
+    }
+  }
+}
