+++ JTime-rest/client/jtime-website/src/store/utils.js
@@ -1,0 +1,1 @@
+export const baseUrl = 'https://jtime.ismail-s.com/api'
